Rename: 상세 페이지 폴더 구조 변경
